FilletBox as a solid-like object instead of a Solid subclass
See examples/fillet-box.py for FilletBox following a "SolidLike"
protocol in order to be treated like a solid by Workplane.

See cadquery/cq.py for an example of the kind of change that Workplane
will need to support SolidLike.  Currently only Workplane.union() has
been adapted.  This is the Workplane method demonstrated in
examples/fillet-box.py.